split derived table tests into their own file
