+++ go/vt/vtgate/semantics/analyzer_test.go
@@ -932,109 +932,6 @@ func TestUnionWithOrderBy(t *testing.T) {
 	assert.Equal(t, TS1, d2)
 }
 
-func TestScopingWDerivedTables(t *testing.T) {
-	queries := []struct {
-		query                string
-		errorMessage         string
-		recursiveExpectation TableSet
-		expectation          TableSet
-	}{
-		{
-			query:                "select id from (select x as id from user) as t",
-			recursiveExpectation: TS0,
-			expectation:          TS1,
-		}, {
-			query:                "select id from (select foo as id from user) as t",
-			recursiveExpectation: TS0,
-			expectation:          TS1,
-		}, {
-			query:                "select id from (select foo as id from (select x as foo from user) as c) as t",
-			recursiveExpectation: TS0,
-			expectation:          TS2,
-		}, {
-			query:                "select t.id from (select foo as id from user) as t",
-			recursiveExpectation: TS0,
-			expectation:          TS1,
-		}, {
-			query:        "select t.id2 from (select foo as id from user) as t",
-			errorMessage: "column 't.id2' not found",
-		}, {
-			query:                "select id from (select 42 as id) as t",
-			recursiveExpectation: T0,
-			expectation:          TS1,
-		}, {
-			query:                "select t.id from (select 42 as id) as t",
-			recursiveExpectation: T0,
-			expectation:          TS1,
-		}, {
-			query:        "select ks.t.id from (select 42 as id) as t",
-			errorMessage: "column 'ks.t.id' not found",
-		}, {
-			query:        "select * from (select id, id from user) as t",
-			errorMessage: "Duplicate column name 'id'",
-		}, {
-			query:                "select t.baz = 1 from (select id as baz from user) as t",
-			expectation:          TS1,
-			recursiveExpectation: TS0,
-		}, {
-			query:                "select t.id from (select * from user, music) as t",
-			expectation:          TS2,
-			recursiveExpectation: MergeTableSets(TS0, TS1),
-		}, {
-			query:                "select t.id from (select * from user, music) as t order by t.id",
-			expectation:          TS2,
-			recursiveExpectation: MergeTableSets(TS0, TS1),
-		}, {
-			query:                "select t.id from (select * from user) as t join user as u on t.id = u.id",
-			expectation:          TS2,
-			recursiveExpectation: TS0,
-		}, {
-			query:                "select t.col1 from t3 ua join (select t1.id, t1.col1 from t1 join t2) as t",
-			expectation:          TS3,
-			recursiveExpectation: TS1,
-		}, {
-			query:        "select uu.test from (select id from t1) uu",
-			errorMessage: "column 'uu.test' not found",
-		}, {
-			query:        "select uu.id from (select id as col from t1) uu",
-			errorMessage: "column 'uu.id' not found",
-		}, {
-			query:        "select uu.id from (select id as col from t1) uu",
-			errorMessage: "column 'uu.id' not found",
-		}, {
-			query:                "select uu.id from (select id from t1) as uu where exists (select * from t2 as uu where uu.id = uu.uid)",
-			expectation:          TS2,
-			recursiveExpectation: TS0,
-		}, {
-			query:                "select 1 from user uu where exists (select 1 from user where exists (select 1 from (select 1 from t1) uu where uu.user_id = uu.id))",
-			expectation:          T0,
-			recursiveExpectation: T0,
-		}}
-	for _, query := range queries {
-		t.Run(query.query, func(t *testing.T) {
-			parse, err := sqlparser.NewTestParser().Parse(query.query)
-			require.NoError(t, err)
-			st, err := Analyze(parse, "user", &FakeSI{
-				Tables: map[string]*vindexes.Table{
-					"t": {Name: sqlparser.NewIdentifierCS("t")},
-				},
-			})
-
-			switch {
-			case query.errorMessage != "" && err != nil:
-				require.EqualError(t, err, query.errorMessage)
-			case query.errorMessage != "":
-				require.EqualError(t, st.NotUnshardedErr, query.errorMessage)
-			default:
-				require.NoError(t, err)
-				sel := parse.(*sqlparser.Select)
-				assert.Equal(t, query.recursiveExpectation, st.RecursiveDeps(extract(sel, 0)), "RecursiveDeps")
-				assert.Equal(t, query.expectation, st.DirectDeps(extract(sel, 0)), "DirectDeps")
-			}
-		})
-	}
-}
-
 func TestScopingWithWITH(t *testing.T) {
 	queries := []struct {
 		query             string
@@ -1179,107 +1076,6 @@ func TestJoinPredicateDependencies(t *testing.T) {
 	}
 }
 
-func TestDerivedTablesOrderClause(t *testing.T) {
-	queries := []struct {
-		query                string
-		recursiveExpectation TableSet
-		expectation          TableSet
-	}{{
-		query:                "select 1 from (select id from user) as t order by id",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select id from (select id from user) as t order by id",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select id from (select id from user) as t order by t.id",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select id as foo from (select id from user) as t order by foo",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select bar from (select id as bar from user) as t order by bar",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select bar as foo from (select id as bar from user) as t order by bar",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select bar as foo from (select id as bar from user) as t order by foo",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select bar as foo from (select id as bar, oo from user) as t order by oo",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}, {
-		query:                "select bar as foo from (select id, oo from user) as t(bar,oo) order by bar",
-		recursiveExpectation: TS0,
-		expectation:          TS1,
-	}}
-	si := &FakeSI{Tables: map[string]*vindexes.Table{"t": {Name: sqlparser.NewIdentifierCS("t")}}}
-	for _, query := range queries {
-		t.Run(query.query, func(t *testing.T) {
-			parse, err := sqlparser.NewTestParser().Parse(query.query)
-			require.NoError(t, err)
-
-			st, err := Analyze(parse, "user", si)
-			require.NoError(t, err)
-
-			sel := parse.(*sqlparser.Select)
-			assert.Equal(t, query.recursiveExpectation, st.RecursiveDeps(sel.OrderBy[0].Expr), "RecursiveDeps")
-			assert.Equal(t, query.expectation, st.DirectDeps(sel.OrderBy[0].Expr), "DirectDeps")
-
-		})
-	}
-}
-
-func TestScopingWComplexDerivedTables(t *testing.T) {
-	queries := []struct {
-		query            string
-		errorMessage     string
-		rightExpectation TableSet
-		leftExpectation  TableSet
-	}{
-		{
-			query:            "select 1 from user uu where exists (select 1 from user where exists (select 1 from (select 1 from t1) uu where uu.user_id = uu.id))",
-			rightExpectation: TS0,
-			leftExpectation:  TS0,
-		},
-		{
-			query:            "select 1 from user.user uu where exists (select 1 from user.user as uu where exists (select 1 from (select 1 from user.t1) uu where uu.user_id = uu.id))",
-			rightExpectation: TS1,
-			leftExpectation:  TS1,
-		},
-	}
-	for _, query := range queries {
-		t.Run(query.query, func(t *testing.T) {
-			parse, err := sqlparser.NewTestParser().Parse(query.query)
-			require.NoError(t, err)
-			st, err := Analyze(parse, "user", &FakeSI{
-				Tables: map[string]*vindexes.Table{
-					"t": {Name: sqlparser.NewIdentifierCS("t")},
-				},
-			})
-			if query.errorMessage != "" {
-				require.EqualError(t, err, query.errorMessage)
-			} else {
-				require.NoError(t, err)
-				sel := parse.(*sqlparser.Select)
-				comparisonExpr := sel.Where.Expr.(*sqlparser.ExistsExpr).Subquery.Select.(*sqlparser.Select).Where.Expr.(*sqlparser.ExistsExpr).Subquery.Select.(*sqlparser.Select).Where.Expr.(*sqlparser.ComparisonExpr)
-				left := comparisonExpr.Left
-				right := comparisonExpr.Right
-				assert.Equal(t, query.leftExpectation, st.RecursiveDeps(left), "Left RecursiveDeps")
-				assert.Equal(t, query.rightExpectation, st.RecursiveDeps(right), "Right RecursiveDeps")
-			}
-		})
-	}
-}
-
 func TestScopingWVindexTables(t *testing.T) {
 	queries := []struct {
 		query                string
@@ -1387,36 +1183,6 @@ func BenchmarkAnalyzeSubQueries(b *testing.B) {
 		"select t.col1, (select (select y.col2 from y) from z) from x as t",
 		"select t.col1, (select (select (select (select w.col2 from w) from x) from y) from z) from x as t",
 		"select t.col1, (select id from t) from x as t",
-	}
-
-	for i := 0; i < b.N; i++ {
-		for _, query := range queries {
-			parse, err := sqlparser.NewTestParser().Parse(query)
-			require.NoError(b, err)
-
-			_, _ = Analyze(parse, "d", fakeSchemaInfo())
-		}
-	}
-}
-
-func BenchmarkAnalyzeDerivedTableQueries(b *testing.B) {
-	queries := []string{
-		"select id from (select x as id from user) as t",
-		"select id from (select foo as id from user) as t",
-		"select id from (select foo as id from (select x as foo from user) as c) as t",
-		"select t.id from (select foo as id from user) as t",
-		"select t.id2 from (select foo as id from user) as t",
-		"select id from (select 42 as id) as t",
-		"select t.id from (select 42 as id) as t",
-		"select ks.t.id from (select 42 as id) as t",
-		"select * from (select id, id from user) as t",
-		"select t.baz = 1 from (select id as baz from user) as t",
-		"select t.id from (select * from user, music) as t",
-		"select t.id from (select * from user, music) as t order by t.id",
-		"select t.id from (select * from user) as t join user as u on t.id = u.id",
-		"select t.col1 from t3 ua join (select t1.id, t1.col1 from t1 join t2) as t",
-		"select uu.id from (select id from t1) as uu where exists (select * from t2 as uu where uu.id = uu.uid)",
-		"select 1 from user uu where exists (select 1 from user where exists (select 1 from (select 1 from t1) uu where uu.user_id = uu.id))",
 	}
 
 	for i := 0; i < b.N; i++ {
